% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnakebiteTools.R
\name{SnakebiteTools}
\alias{SnakebiteTools}
\title{SnakebiteTools Package}
\description{
Collection of Tools to Analyse and Monitor Snakemake Runs
}
\author{
Daniel Fischer \email{daniel.fischer@luke.fi}
}
