% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake_log_analyser.R
\name{snakemake_log_analyser}
\alias{snakemake_log_analyser}
\title{Analyse Snakemake runs}
\usage{
snakemake_log_analyser(project_folder, print_sacct = TRUE, no_return = TRUE)
}
\arguments{
\item{project_folder}{Location of your project}

\item{no_return}{logical, turn off the return value of the function}

\item{print_feedback}{logical, should the sacct command be printed}
}
\value{
A data.frame with the job information from the latest log file
}
\description{
Get more inforamtion about the status of your snakemake run
}
