% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake_log_analyser.R
\name{snakemake_log_analyser}
\alias{snakemake_log_analyser}
\title{Analyse Snakemake runs}
\usage{
snakemake_log_analyser(project_folder)
}
\arguments{
\item{project_folder}{Location of your project}
}
\value{
Feedback and, if applicable a sacct command for further information
}
\description{
Get more inforamtion about the status of your snakemake run
}
